+++ celery/tornium_celery/celery_app.py
@@ -178,7 +178,6 @@ if celery_app is None:
             "tasks.guild",
             "tasks.items",
             "tasks.misc",
-            "tasks.reports",
             "tasks.stakeout_hooks",
             "tasks.stocks",
             "tasks.user",